[16b1cba] [vfs] IFile::Read and all derived classes: change to use size_t/ssize_t, do not use more than SSIZE_MAX of buffer for "Read()", small related fixes

+++ xbmc/filesystem/FileReaderFile.cpp
@@ -64,7 +64,7 @@ bool CFileFileReader::OpenForWrite(const CURL& url, bool bOverWrite)
 }
 
 //*********************************************************************************************
-unsigned int CFileFileReader::Read(void *lpBuf, int64_t uiBufSize)
+ssize_t CFileFileReader::Read(void *lpBuf, size_t uiBufSize)
 {
   return m_reader.Read(lpBuf,uiBufSize);
 }
